+++ test/grinder_utils_test.dart
@@ -8,12 +8,32 @@ import 'dart:io';
 import 'package:grinder/grinder.dart';
 import 'package:unittest/unittest.dart';
 
+import 'utils.dart';
+
 main() {
   group('grinder.utils', () {
     test('get sdkDir', () {
       if (Platform.environment['DART_SDK'] != null) {
         expect(sdkDir, isNotNull);
       }
+    });
+
+    test('pub version', () {
+      MockGrinderContext context = new MockGrinderContext();
+      Pub.version(context);
+      expect(context.isFailed, false);
+    });
+
+    test('dart2js version', () {
+      MockGrinderContext context = new MockGrinderContext();
+      Dart2js.version(context);
+      expect(context.isFailed, false);
+    });
+
+    test('analyzer version', () {
+      MockGrinderContext context = new MockGrinderContext();
+      Analyzer.version(context);
+      expect(context.isFailed, false);
     });
   });
 }