+++ changelog.md
@@ -22,6 +22,9 @@ PubTool's build methods now take an optional `workingDirectory` argument.
 Remove `runSdkBinary` and `runSdkBinaryAsync`, and they are no longer needed.
 Use `runProcess` and `runProcessAsync` instead.
 
+The methods on `PubTools` and `Dart2jsTools` are now static - you no longer need
+to create an instance to use them.
+
 ## 0.5.7 (2014/07/28)
 
 - added `runProcessAsync()` and related async methods (such as `PubTools.buildAsync(...)`)