+++ changelog.md
@@ -23,7 +23,8 @@ Remove `runSdkBinary` and `runSdkBinaryAsync`, and they are no longer needed.
 Use `runProcess` and `runProcessAsync` instead.
 
 The methods on `PubTools` and `Dart2jsTools` are now static - you no longer need
-to create an instance to use them.
+to create an instance to use them. `PubTools` renamed to `Pub` and `Dart2jsTools`
+renamed to `Dart2js`. A new utility class for dartanalyzer - `Analyzer` created.
 
 ## 0.5.7 (2014/07/28)